Change Product Suggestions Tests to Assert Eventually to handle elastic search delay time
