+++ commercetools.Sdk/IntegrationTests/commercetools.Sdk.IntegrationTests/TestingUtility.cs
@@ -385,42 +385,5 @@ namespace commercetools.Sdk.IntegrationTests
         }
 
         #endregion
-
-        public static void AssertEventually(TimeSpan maxWaitTime, TimeSpan waitBeforeRetry, Action runnableBlock)
-        {
-            long timeOutAt = (int) DateTime.Now.TimeOfDay.TotalMilliseconds + (int) maxWaitTime.TotalMilliseconds;
-            while (true)
-            {
-                try
-                {
-                    runnableBlock.Invoke();
-                    // the block executed without throwing an exception, return
-                    return;
-                }
-                catch (Exception ex)
-                {
-                    if ((int) DateTime.Now.TimeOfDay.TotalMilliseconds > timeOutAt) //if it's timeout
-                    {
-                        throw;
-                    }
-
-                    if (ex is ErrorResponseException errorResponseException &&
-                        errorResponseException.ErrorResponse.Errors.Any(err =>
-                            err.Code.Equals("SearchFacetPathNotFound")))
-                    {
-                        throw;
-                    }
-                }
-
-                try
-                {
-                    Task.Delay((int) waitBeforeRetry.TotalMilliseconds).Wait();
-                }
-                catch (ThreadInterruptedException e)
-                {
-                    throw new SystemException(e.Message);
-                }
-            }
-        }
     }
 }
